app/requirements.txt
flask
redis
paho-mqtt
